requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic[email]==2.10.3
email-validator==2.2.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
python-dotenv==1.0.1
